int > np.int64 in `dtype_integer`

+++ geometric_kernels/lab_extras/numpy/extras.py
@@ -73,4 +73,4 @@ def dtype_integer(reference: B.NPRandomState):  # type: ignore
     """
     Return `int` dtype of a backend based on the reference.
     """
-    return int
+    return np.int64
